vdpau: Improve detection of supported files.
A minor improvement on the detection of whether or not a video will
play. It is a good idea to at least make sure that the video codec
is supported and that the resolution does not exceed the supported
resolution.

v2: Changed detection of errors to make use of CheckStatus.
v3: removed tests for the definition of VDP_DECODER_PROFILE_MPEG4_PART2_ASP