vdpau: Improve detection of supported files.
A minor improvement on the detection of whether or not a video will
play. It is a good idea to at least make sure that the video codec
is supported and that the resolution does not exceed the supported
resolution.

V2: Added suggested changes to use CheckStatus instead of directly
    testing the status.